Add initial support for 27.1 in app-toolkit, switch Paging to use it
Bug: 72443497
Test: tests in paging:runtime, paging:common, room:integration-tests:*
Test: ./gradlew createFlatfootDocsArchive -PofflineDocs=true --stacktrace --debug
Test: ./gradlew runBuildServerCompilationTasks

In supportlib 27.1.0:
    ListAdapter moved to v7-recyclerview
    ListAdapterHelper moved to v7-recyclerview as AsyncListDiffer
    ListAdapterConfig moved to v7-recyclerview as AsyncDifferConfig
    DiffCallback moved to v7-recyclerview as DiffUtil.ItemCallback

This CL updates paging to account for those moves. Because of this,
previous versions of paging cannot be used with 27.1, and future
versions cannot be used with v7-recyclerview < 27.1.

Switched setList->submitList throughout to match new naming.
AsyncPagedListDiffer can't easily implement a @NonNull
getCurrentList() like AsyncListDiffer does, so leaving the rest of
its API as is.

Only paging and tests that depend upon it use 27.1 for now, other
projects must opt in.

Preemptively sorted out kotlintestapp's dependencies to make it
compatible with 27.1, even though it doesn't use it yet.

Change-Id: Ia43203a51b82cdda68b4bf503a1166ab254a1aa3