Wrote an implementation of R4A that uses the IR
Our old code generation strategy was to introduce extension points that allowed us to poke
at the compiler internals and effectively "trick" it into producing roughly what we wanted.  Then
we could go in after the fact and massage the bytecode until we got the desired output.  This
strategy was challenging to write/maintain and very error prone, leading to subtle bugs.

This commit is effectively a rewrite of the entire R4A compiler plugin to use the new IR backend.
This new approach will allow us to utilize the IR pipeline to perform transformations prior to
bytecode generation, which will provide us with a much better foundation for R4A's future.

This code still needs a bunch of cleanup, but I figured it's better to get it merged now
so we everyone else can start experimenting on top of it, rather than to wait until it's
perfect.  I'll go ahead and submit some cleanup diffs soon.

Change-Id: I0bb9218d0be22859c3b12d03b4d91c961c8aa9ce