Move metalava setup out of init.gradle
Prior to this we were just copying what Doclava was doing but there is no reason to do that. *.gradle scripts are legacy and we should focus on moving everything to the Kotlin code.

This also restores Metalava to working order. Its prebuilt was removed, but its tasks were not correctly set up to run on CI. This change only restores Metalava's tasks to working order. A follow-up change will ensure the tasks are run as part of the buildOnServer task.

Test: ./gradlew :core-ktx:checkApi
Change-Id: Id6dd21a4145b3b7dd4d1f8a90b48f3d086ca2b48