Add support for variable capture and parameters in KTX bodies.
Additionally, parses a KTX body as a FUNCTION_LITERAL,
which should presumably make typechecking of lambda perameters easier.
To be determined.  I think Leland said he had some type resolution stuff
he was rewriting, so there is a bit of a chicken-and-egg problem where
we each want what the other person has, so we might as well start merging
the stuff.

Change-Id: I24fa8eae89e91444d924774064d71ad530fc6202