Ask Gradle to prefer project modules
When Gradle finds an androidx external module and its local
project to have the same version, it prefers the external module
over the local version. This could have been fine except that
Gradle seems to still pick an alpha01 external module over an
alpha02 project (rest of the version being the same) which makes
us think it's probably a bug Gradle or it just doesn't support
alpha and beta stuff properly... So to avoid that, choose prefer
project instead.

Test: ./gradlew dependencies and find that projects are
preferred over same version external modules and that
it does not substitute modules that should not be
substituted (i.e it does not pick the local project as
dependency if there is no conflict to begin with.)

Change-Id: I45bd67bf279d5852c8bb8b1c0cdff1db01006e88