Fix pom change detection in Jetifier.
Jetifier is not supposed to generate a new archive if nothing has
changed (did not have any support library reference). However that was
not the case if archive contained a pom file as the detection was broken
and it always reported change. This is mostly significant when a library
depends on a signature and we modify it. We also had a test for it
already, but there was a bug in the test itself :( So the test was
asserting incorrect state.

I have also manually verified this on independent library and can
confirm that we no longer trigger change.

Test: Fixed :(
Bug: b/118785891
Bug: b/122720498
Change-Id: Ie53cbc3f4b0d78280913a0867584eaba737e8929