Move core to 1.2.0-alpha02
Pin all usages of ToT core where the version > alpha
Version 1.2.0-alpha01 was released on May 7th

Note: Pinning the version of core in appcompat caused lint to throw
a wrong constant error.  Jeff and I couldn't really figure out
why, so we're suppressing for the time being.  It's interesting
that pinning a dependency caused this error to occur.

Test: ./gradlew verifyDependencyVersion checkApi && ./gradlew bOS
Change-Id: I6b366f315429138d6aa5746052c9522f9d0a8e2f