Introduce OpaqueKey
Currently all the internal helper keys we use inside compose-runtime are implemented like this
private val RootKey = Any()
It works fine now, but we will want to support calculating of the compound keys(combination of hashes for all the keys of currently open groups). And we also want this key to stay the same when we restart the app after the process kill. With the current implementation Any().hashCode() is not equals Any().hashCode() which will be executed after the process recreation. To have fixed hashCodes() for such keys we introduce OpaqueKey - simple data class with only one field - String key. Now we will write the same key like this:
private val RootKey = OpaqueKey("root")
With such implementation RootKey.hashCode() will stay the same.

Bug: 150453289
Test: manually
Change-Id: I19f895af19df8288818b3922f0c93e97bf612ac9