Update references to inner class constructors.
Previously we had to have fully qualified names to ctors,
despite them being invalid javadoc. The other CL in this topic
fixes the underlying doclava bug so we can write proper javadoc now.

Due to how the public docs are generated (from prebuilt srcjars),
added a flag to the publicDocsTask to ignore the errors. This part
of the change will be reverted once new prebuilts can be generated
with the proper javadoc syntax.

Bug: 6963924
Test: ./gradlew disttipOfTreeDocs
Change-Id: I62672894e8d7bfb42e17d5accb20e8721e224db8