Update POM URL to point to specific release
Currently, the POM URL points to generic Jetpack or AndroidX homepages. This CL points the POM URL to the specific release in question, for example, https://developer.android.com/jetpack/androidx/releases/fragment#1.3.0-alpha06. This will show up on gmaven and in the version updater plugins as a helpful way to get release notes.

Test: manually ran createArchive and checked the outputted URLs.

Change-Id: I813f3d83dce0f319c0a905d24f60f09aa5b90f8d