Use MonotonicFrameClock as AnimationClockObservable
Introduces an AnimationClockObservable that takes a MonotonicFrameClock,
so you can use a MonotonicFrameClock in the current animation APIs.

Moved one ScrollableTest to MonotonicFrameClock as a proof of concept.

Bug: 161247083
Test: ./gradlew compose:foundation:foundation:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.compose.foundation.ScrollableTest#scrollable_horizontalScroll
Relnote: "Added MonotonicFrameAnimationClock that enables you to use a
MonotonicFrameClock as an AnimationClockObservable to bridge the gap
between the new coroutines based clocks and APIs that still use the old
callback based clocks.

The MonotonicFrameClock equivalent of ManualAnimationClock is now
ManualFrameClock."

Change-Id: I111c7b7182a1495f95eab1bb808d3acd6af82447