Add InteractionState to scrollable and corresponding containers, add animation observation to LazyList
This CL adds InteractionState to the ScrollableColumn/Row and Lazy lists to allow people to observe interactions that are happening on this element. While doing so, to make functionality on par, I added isAnimationRunning and stopAnimation is LazyState, so people won't miss this menthods while migrating from ScrollableColumn.

Change-Id: I815666e1b4544dcd5da9f253ec1b539fdd777529
Relnote: Added interactionState to Modifier.scrollable, ScrollableColumn and LazyColumnFor
Fixes: 169509805
Test: added