Introduce GraphicsLayerScope and a new Modifier.drawLayer overload
This overload allows to provide a lambda block on GraphicsLayerScope where you define the layer parameters in a way which allows to skip recomposition and relayout when the state change happens. DrawLayerModifier is now internal in preparation to migrating its logic into placeable.placeWithLayer() method of LayoutModifier

Relnote: Added a new Modifier.drawLayer() overload. It takes a lambda block on a new GraphicsLayerScope where you define the layer parameters in a way which allows to skip recomposition and relayout when the state change happens. DrawLayerModifier is now internal in preparation to migrating its logic into placeable.placeWithLayer() method of LayoutModifier
Test: ./gradlew bOs
Bug: 173030831
Change-Id: I15e9f41e3c93245529bf798e4a9617ca49d5e509