Use RestorableStateHolder in lazy lists
RestorableStateHolder allows to save the state defined with [savedInstanceState] and [rememberSavedInstanceState] for the subtree before disposing it to make it possible to compose it back next time with the restored state. We use it to save the state for items which are no longer visible in LazyColumn to be able to restore item's state like inner LazyRow scroll position when the item with LazyRow would became visible again.

Bug: 166586419
Test: new demo, new tests in LazyItemStateRestoration
Change-Id: I595a4eb517786b4a4aa30bf65c91399b3de0a510