Add the intrisic ability to build trees bottom-up.
Before this change the, to build a tree bottom up, the Applier
needed to maintain its own list of pending inserts and apply
them during during the `up()` method when leaving the node.

With this change the `insert()` method has changed to be
`insertTopDown()` and a new method, `insertBottomUp()` was
added to allow an applier to override one or the other depending
which is more efficient.

Fixes: b/174173910
Test: Existing tests refactored and pass

Relnote: """The `Applier` interface has changed to simplify
building trees bottom-up instead of top-down.

The `insert()` method has been renamed to `insertTopDown()`.

A new method, `insertBottomUp()`, was added.

An applier either inserts nodes into the tree it is editing
using `insertTopDown()` or `insertBottomUp()` depending on
which performs better.

Some trees, such as `LayoutNode` and `View`, are much more
efficient to build bottom-up than top-down. Prior to this change,
a stack of inserts was required to implement bottom-up which
needed to be coplied to every applier which needed bottom-up
construction for performance. With this change an `Applier`
overrides `insertBottomUp()` to build a tree bottom-up and
`insertTopDown()` to build the tree top-down.
"""

Change-Id: Icbdc2929ab8fc8fce231d633b062fc80be5c10c9