Support nested scroll for backdrop, modal sheet, expanded sheet and bottom drawer.
This CL supports nested scroll by opening a few APIs in swipeable state, so any user that uses it would be able to set their desiged behaviour.

And them it implements such behaviour for main swipeable components that need it.

Note that swipeable is swiping, but not scrolling, therefore it doesn't dispatch the scroll events, only participates as the connection.

Test: Added new tests to test NestedScrollConnection in swipeable + demos
Bug: 162408885
Change-Id: Iaa17a40fec2949f6364a9339a5d269dc622570d6
Relnote: added API to manually trigger settle animation and drag in Modifier.swipeable