Add new test-suite-tag for media compat tests
This gives us the option of running these media tests
either on their own, or as part of regular runs. Added
because media currently tests against a superset of the
devices we use for all libraries, but I don't believe the
additional devices can support running all of our tests.

Bug: 151744308
Test: via forrest: https://g3c.corp.google.com/results/invocations/5dbfcd40-b67d-4eeb-abfc-099366dcf633/targets
Change-Id: I6fa90cc19232f3de1bc37c76af291c2ff733acf4