Changes Tab and BottomNavigationItem to use nullable parameters instead of parameters defaulting to emptyContent()
In these components, the overall size and layout changes depending on whether there is content in these parameters or not, and it is confusing to represent this implicitly, by checking if there is a layout or not inside the composable lambda. Instead we can use `null` to represent this semantic difference.

Fixes: b/152764416
Test: TabTest
Test: BottomNavigationTest
Relnote: "Changed Tab's `text` and `icon` parameters, and BottomNavigationItem's `label` parameter to be nullable, to better convey the behavior of the component when these parameters are / are not provided, since it affects the size and layout of the component. If you are currently passing `emptyContent()` to represent no text / icon / label, you should use `null` instead."
Change-Id: I57ed4f72c434440068ab17edfa2b5ef791ddbf2f