Allow nullable action in AccessibilityAction
If multiple semantics blocks with the same AccessibilityAction are
provided, the collapsed/merged AccessibilityAction will be
AccessibilityAction(outer.label ?: inner.label, outer.action ?: inner.action),
where outer is the AccessibilityAction in the outer semantics block
and inner is the AccessibilityAction in the inner semantics block.
This allows developer to provide a different action label to a built-in
semantics action.

Note custom actions are still non-null because custom actions doesn't have
a separate key, so we can't tell which two actions are the same action.

Relnote: "Allow nullable action in AccessibilityAction, change action
label in AccessibilityAction and CustomAccessibilityAction from
CharSequence to String"

Fix: b/168033885

Test: tested manully with IconToggleButton in the demo app with talkback
on. Unit tests added.

Change-Id: I0dc82f11f78ebe89be161727feed53e1c2390b72