Move TextDelegate to foundation
TextDelegate is an internal class mainly used by
foundation group. Since groups cannot depend on
internal APIs, moved TextDelegate up to foundation.

When we are confident about what kind of API we want
for TextDelegate (including the name) we can
reposition the class in the related module (possibly
Ui-text)


RelNote: N/A
Test: Treehugger
Bug: 179710063
Change-Id: Id77d6189d958c7b1c2fb7fcfebea8e3c5b99eec4