Fix taking nested snapshots with diposed parents
Creating a nested snapshot from a nested snapshot delegated
the create to the parent snapshot which would throw if it
was applied or disposed.

This change creates the new snapshot directly instead as it was
incorrect to validate the parent was still valid and second, it
was incorrect take the snapshot from the current original parent
current state instead of the snapshot state being snapshot.

Fixes: b/181159260
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I74bc1a11e85cf9b5e7e4bd830913c5d4d9dcdb5d