Add scrim semantics in Drawer and ModalBottomSheet
When the drawer is open or modal bottom sheet is visible, we have a
scrim with a tap gesture filter. This tap gesture prevents touches to
contents behind them. We also report this tap gesture to semantics so
that accessibility oclusion design for completely covered sibiling will
work properly.

Also added SemanticsNodeWithAdjustBounds to record adjusted
bounds after the oclusion and use it for accessibility nodes. The
effect will be like:
https://hsv.googleplex.com/6415050837327872

Note this use scrim node for closing drawer in accessibility is an
approved design.

Also when we get all the current semantics nodes, we only includes
nodes that are placed to avoid errors of asking bounds for
unmeasured/unplaced nodes.

Fix: b/153643868, b/153198816
Bug: b/153644956

Test: manually tested with drawer and modal bottom sheet with talkback.
Unit tests added.

Change-Id: I6dd7e4615615b464d8dc9afde35261cb37836c23