Fixes ripple being stuck if cancelled immediately
Because we only added the ripple to the list of ripples to draw inside a coroutine, it's possible that the ripple could be removed before it was added to the list.

Now we synchronously add the ripple, and launch animations separately to avoid this problem.

Additionally fixes not animating the state layer in a separate coroutine - in both cases it is important that we launch the animations separately so that we don't wait for the animation to finish before we react to new Interactions, as we might need to cancel / change target mid-animation.

Fixes: b/181253315
Test: Existing tests
Test: Demos
Change-Id: I6a13ddf9b2ef9aee03a5fb3ac01decaa5e197ca3