Move wear:wear-tiles(renderer,proto) to wear:tiles:tiles(renderer,proto)
This allows us to treat wear:tiles as a same version group. The main
advantage here is that we don't have to track the surface of
tiles-proto, since it's an auto-gen'd set of classes that we can't
easily add custom annotations to.

Test: N/A - refactoring.
Relnote: Moved wear:wear-tiles* artifacts to wear:tiles:tiles*.
Change-Id: I824eafa414aa6ac8957c1ffd209ae11612a87365