Fix exported activities in manifest files
ComponentActivity in ui-test-manifest is now exported, and the
ComponentActivity with TestTheme is now exposed to all compose libraries
through test-utils instead of through ui-test and ui-test-junit4.

Also fixes the TestActivity from ui-tooling that was inadvertently
exported as well.

Bug: 184718994
Test: ./gradlew cC
Relnote: "The AndroidManifest files from ui-test-manifest and
ui-tooling-data are now compatible with Android 12"

Change-Id: I6f9dec0515ad6eb7fd232eeb124ee0164f4e90cb