tryAwaitRelease in detectTapGesutres now returns proper value.
Since the migration to channels in detectTapGestures, we cannot reset anymore at the start of the loop, since the coroutine resume order is not guaranteed, as we have a separately launched loop that doesn't need to wait for all up. We don't have to reset at the start of the loop either, since the time we care about press is after down, so the rest logic moved there.

Fixes: 184872415
Test: added regression test
Change-Id: I2cd3a6946ecf6fd565d4b18681729cf6e3c4ae96