Improve how CornerBasedShape handles oversized corner sizes
Previous logic was giving all the available size for the top corners which was proved to be unpredictable. The new logic proportionally scales both top and bottom corners if their sum is larger then the available size.
Note that we do not just clamp corner sizes to half of the corresponding dimension in order to support a perfectly valid use case RoundedCornerSize(topStart = CornerSize(percent = 100), bottomStart = CornerSize(percent = 0))

Test: CornerBasedShapeTest
Fixes: 181908570
Change-Id: I44768ce322b0b0cd06f6683dcbbc52c8e1d8126e