Apply background color in OutlinedTextField too
It's a bit tricky how we should draw background when there is a Label
(i.e. include padding or not), but in any case it will look weird when text field is focused.
But when there is no Label, it makes total sense to change the background of
the OutlinedTextField. Therefore I propagate background color from the
defaults to the implementation.

Test: new test in OutlinedTextFieldTest
Fixes: 184935344
Change-Id: Ib13fc9176b4f96a20a246f9860b2d3dca1c2af58