Revert "Make scrollable to adjust velocity if some delta consumed during "pre"-stage of the nested scroll"
This reverts commit df1da916ff4c7fc0acff00e8988923f53aa6e0d2.

Reason for revert: It's a harmful fix, rendering nested scroll unusable for cases when the parent consumes all the delta all the time. This change also breaks the geometry of velocity as a whole, so the velocity is not following the finger physics.

Relnote: n/a
Test: n/a
Change-Id: I23de94bba5e58d1c04cfed2b653788005d6bf968
(cherry picked from commit c235569e0f8dee12ab42657d262b1988fddb443b)