Ignore AnimatedVisibility animations in AnimationPreview
AnimatedVisibility animations are being picked up by the Animation Preview
as if they were a Boolean animation. The problem is these animations cannot
be controlled from the Animation Preview, since who actually animates is a
child transition, which is not something supported by Animation Preview yet.

We need to ignore AnimatedVisibility animations for now so they are not
displayed at all in the Animation Preview, otherwise extra tabs will be
shown to the users and that only adds an extra layer of confusion, since it
deviates their attention from the tabs representing actual transitions.

Bug: 193807833
Test: ComposeViewAdapterTest#animatedVisibilityIsIgnored
Change-Id: I243a3b775312c6a3a41eb5026ac649a2700fb554