Allow state to move within a composition
Added `movableContentOf` which converts a composable lambda
into a lambda that moves the state from the last place
it was called to the new place it is called.

Added `movableContentWithReceiverOf` which is identitcal to
`movableContentOf` but allows creating a lambda with a
receiver parameter.

Fixes: 168929878
Test: ./gradlew :compsoe:r:r:tDUT

Relnote: """Added `movableContentnOf` which converts a composable
lambda into a lambda that moves it state, and corresponding nodes,
to the any new location it is called. When the previous call leaves
the composition the state is temporarily preserved and if a new call
to the lambda enters the composition then the state, and associated
nodes, are moved to the location of the new call. If no new call is
added the state is removed permantently and remember observers are
notified.

If a `movableContentnOf` lambda is called multiple times in the same
composition, new state and nodes are created for each call and, as
calls leave the composition and new calls enter, the state is moved
from the first leaving calls to the entering calls in the order they
are called. All state not claimed by new calls is removed
premanently."""

Change-Id: Ib4850095f241617a191ea7815fc947adaf867456