Redo the implementation of inspectable
The foldIn and foldOut logic did not work well when combined with
composed modifiers which implementation is calling foldIn repeatedly.

Instead use a start and end marker in the list of Modifier elements.

This fixes a set of tests that would fail with the earlier
implementation.

Bug: 191017532
Test: Fixes existing tests after using inspectable
Relnote: Redo implementation of inspectable
Change-Id: I927bc731712d81f47aefdf946f778047c7645b26