Desktop Accessibility
This CL contains initial support of Accessibility. We use the abilities of AWT to support different operating systems via uniform API. The CL contains main infrastructure elements: the passing of our implementations to AWT, synchronization of semantics nodes with AWT-representations, translation of main SemanticProperties to corresponding a11y roles and properties, and also sending messages about some property changes. Many things are not working as they should, but this is the beginning.

Test: manual
Change-Id: I3b628a57673f3b9331cd07dcfcaaa5e6fcc41f03