Introduce LazyListPositionedItem
Convert LazyMeasuredItem into an immutable LazyListPositionedItem when we know the offset. Previously we had mutable LazyMeasuredItem which gets it offset updated later on which is not safe. We were even using this exact object as LazyListItemInfo which assumes the object is finalized. Now we have an extra step which is a wrapper on top of placeables which already know their intended positions. It will also simplify the logic handling the item position animations

Test: tests in lazy package
Change-Id: I458f38a71ab31fde44aa71b3ac9550b90b595ce8