Supply an invariant id to InspectorNodes
Add an identity property to CompositionGroup, which creates an
anchor and returns the instance without type (Any).

For every CallGroup that is expected to be converted into an
InspectorNode, create an anchor from the CompositionGroup.

The LayoutInspectorTree computes an id from the anchor instance
by converting the hashcode of the anchor to a negative id value.

This id value should be unchanged after recompositions which will
allow the Layout Inspector UI to re-stablish open nodes in the
component tree.

Fixes: 199315013
Test: LayoutInspectorTreeTest.testCrossfade
Relnote: "Added identity field to CompositionData for generating
invariant ids in the Layout Inspector."

Change-Id: Ic116e6682233c31ccc4a81f8cf0cc96cac4a83ca