Don't expose receiver scopes for touch and mouse injection
Instead of exposing `Touch: TouchInjectionScope` and
`Mouse: MouseInjectionScope` in MultiModalInjectionScope, expose a
function for each modality that accepts a lambda with the appropriate
receiver scope. This prevents exposing the receiver scope directly.

Fix: 204078403
Test: ./gradlew compose:ui:ui-test:cC
Relnote: "Removed `MultiModalInjectionScope.Touch` and
`MultiModalInjectionScope.Mouse`. In order to inject touch and mouse
events for multi-modal gestures, you can now use
`MultiModalInjectionScope.touch()` and
`MultiModalInjectionScope.mouse()`, both of which accept a lambda that
has the receiver scope of that modality."

Change-Id: Idde18dc95d6b71c8ac15772198ef5de8b612a5b6