Remove TestComposeWindow in favor to ImageComposeScene
This CL introduces ImageComposeScene - simplified version of ComposeScene, with drawing only into Image.

It is almost the same as the old TestComposeWindow, but with a different class and method names. TestComposeWindow was deprecated before Compose 1.0, so it is safe to remove.

See examples of usage in ImageComposeSceneTest.

Also, we introduce:
- renderComposeScene - one-liner function to draw Composable content into an image
- ImageComposeScene.use - similar to AutoClosable.use to disposing created resources. We don't inherit AutoCloseable, because we want to move ImageComposeScene into skikoMain. For now, common code doesn't have Closeable, but in the future it can (https://github.com/Kotlin/kotlinx-io/blob/c5bba0114e33c2afd8e4b1ab214aad94ee416d00/core/commonMain/src/kotlinx/io/Closeable.common.kt#L7), so I renamed `dispose` method to `close.

ImageComposeScene still has some flaws, which will be fixed in the future (https://github.com/JetBrains/compose-jb/search?q=ImageComposeScene&type=issues)

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Change-Id: I48d2769fd2eaff53e391371aea9ed37964b6afc3