Introduce `isSkipHalfExpanded` flag for Modal Bottom Sheets
If the sheet is tall enough to have a half-expanded state, this state can be disabled by passing `true` for `isSkipHalfExpanded` or by setting `ModalBottomSheetState`'s `isSkipHalfExpanded` property to `true`.

Updated the `ModalBottomSheetSample` to have a toggle for this setting.

Relnote: `ModalBottomSheetState` now has a `isSkipHalfExpanded` flag. It can be either set through the constructor or updated later on by setting `ModalBottomSheetState`'s `isSkipHalfExpanded` property to `true`. Updating `isSkipHalfExpanded`'s value causes a recomposition of the sheet.

Bug: b/186669820

Test: `modalBottomSheet_showAndHide_manually_skipHalfExpanded`, `modalBottomSheet_hideBySwiping`, `ModalBottomSheetStateTest`
Change-Id: I18b866d1895f53d46c5965867f628b5eca455fa4