Adds lint check to warn when using mutableStateOf with mutable collections
Mutating a mutable collection inside a mutable state will not cause a recomposition, as the mutable state is not aware of changes to objects inside. Instead, you should either use a read-only collection type such as List, and assign a new instance to the mutable state when your data changes, or use a snapshot-backed collection type such as mutableStateListOf / mutableStateMapOf that will observe mutations to the collection.

Fixes: b/205985838
Test: MutableCollectionMutableStateDetectorTest
No-Typo-Check: 'Typo' in generated hash
Change-Id: I3df3a7528ec9a913732cbff4e69d28bd3cb34258