Desktop. Fix race when we call ComposeScene.render from another thread
The race was encountered when we switched from TestComposeWindow to ImageComposeWindow.

The race was between:

[Swing thread]
(1) GlobalSnapshotManager -> sendApplyNotifications
(4) parentFrameClock.withFrameNanos [we stuck here, because there is no frame after that]

[Test thread]
(2) ComposeScene.render -> sendApplyNotifications [we don't apply anything, because of (1)]
(3) frameClock.sendFrame

To make sure that recomposer will wait for the frame, we had dispatcher.flush.
But it don't wait for the already running tasks (in our case - for sendApplyNotifications to finish in Swing thread).
In this CL we fix that, adding a synchronization point - we don't exit flush if there are still performing tasks.

Also, we use FlushCoroutineDispatcher only for recomposition now, and don't use it for effects to avoid unpredictable behaviour on the user side.

We still have races between GlobalSnapshotManager and ComposeScene if ComposeScene is used outside of the Swing thread (see https://github.com/JetBrains/compose-jb/search?q=ImageComposeScene&type=issues). I will fix them later.

Also, I changed default dispatcher from EmptyDispatcher to Unconfined. Empty dispatcher has no use, if we want to test animations.

Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true
Change-Id: I2fc4c5524e6f71d530b0dbce4a99fb5bf1c15110