Fix distance check for minimum touch target
Fixes: 223588290

Previously, the minimum distance in the horizontal or vertical
direction from the layout bounds was used to determine the
closest touch target when in the minimum touch target area.
This causes a strange effect where you can be right next to
one layout and farther from another layout and the farther
one could be selected.

This CL uses the actual distance from the layout bounds
to determine which is closer rather than treating each
axis separately.

Test: new test
Change-Id: I530714f48fab0328bfcfb011690f5d7017916632