API Council Updates: rememberNestedScrollConnection
In this CL I have updated the NestedScrollConnection to use a generic View instead of a specific ComposeView. I have also acquired the view from the composition locals LocalView.current, making the API easier and more aligned with other APIs in the framework.

Test: All previous tests should pass.
Relnote: "Simplified the rememberNestedScrollConnection API to use composition locals to acquire the current view information"

Fixes: 229621931

Change-Id: I67ca70fdf69493105ec9487fd5f1f44996d9619e