Introduce Modifier.Node APIs, refactor Modifier/Wrapper/Entity strategy
Bug: 239955664
Test: Added tests, and also most existing tests exercise this code
Relnote:“””
The layer which handles Modifier/Modifier.Element instances and coordinates their behavior on LayoutNodes has been majorly refactored.

As it stands this was a refactor which did not affect the public API of any of the many modifiers in Compose, and can be viewed as an implementation-only change. Despite that, this is an important change for various reasons.

Change Summary
===
The added experimental Modifier.Node APIs provide an abstraction that allows for state to be maintained on an instance that will be retained with the lifecycle of the layout node, and will be allocated per-layout-node and per-usage of the corresponding Modifier.Element that produced it.

Broadly speaking, this abstraction provides an alternative mechanism to produce stateful modifiers without relying on the mechanics of the Modifier.composed API.

Risk
===
This change is strictly binary compatible with prior releases, and is intended to be backwards compatible in terms of observable behavior as much as practical and reasonable. That said, there are few subsystems of compose this refactor did not touch, and *it is likely that behavior has changed in ways that were not covered by our tests and have not yet been found and fixed*.

Please upgrade to this release with caution. If you believe this has broken something for you, please let us know.

Experimental APIs
===
Various experimental APIs have been added, all relating to the new concept of a “Modifier Node”. Modifier.Node’s are created as a result of

fun modifierElementOf(…): Modifier
abstract class ModifierNodeElement
abstract class Modifier.Node
abstract class DelegatingNode
interface LayoutModifierNode
interface DrawModifierNode
interface SemanticsNode
interface PointerInputNode
interface ModifierLocalNode
interface ParentDataModifierNode
interface LayoutAwareModifierNode
interface GlobalPositionAwareModifierNode
interface IntermediateLayoutModifierNode
“””

Change-Id: Ie4313071d4d9ebc82b28b6fbb691d3736d7d738a