Include list of files & paths for a given artifact
This is the next CL in a series towards displaying artifact ID in reference docs. Note that the flag for showing the metadata (`showLibraryMetadata`) is still set to false, so this feature is a no-op in terms of generated refdocs.

The list of files & path will be used by Dackka to look up a class's
artifact ID.

Bug: 232141800
Test: Remove `@Transient` from `jarContents`, run `./gradlew  :docs-public:generateMetadata`, and verify the contents in the generated JSON file (in out/androidx/docs-public/build) contain a list of files for each metadata entry. Set `DEBUG` to `true` to pretty-print the JSON file
Change-Id: If67f089fb32287ba32316c77170731014761c901