Restructure main page
Extract tutorials and example to separate sections. 

Also add links to Android documentation. This documentation is very comprehensive comparing to our current documentation, that covers mainly desktop aspects.

Until we write our own docs, or unify Jetpack Compose docs with Google collaboration, it is better to have links to this documentation than not to have it.