Fix crash in SemanticsSort
Found a rare reproduce that caused this crash. Returning -1 from
compareTo() is not correct, even if input list was sorted by z-order.
The fix is to break a tie, for example by using semanticsId, similarly
to how it's done in ViewGroup.ViewLocationHolder.compareTo().
Also added z-order comparison to preserve accessibility focus order.

Fixes: 207477257
Test: ./gradlew :compose:ui:ui:testDebugUnitTest
Change-Id: I6ebb5e9b16f1d2a1b0239ee9178c542b10f52d9b