Fix idea plugin compatability (#2318)
IDEA doesn't provide backward compatibility policy, and breaking changes can happen in any version, without prior deprecation:

https://plugins.jetbrains.com/docs/intellij/api-changes-list.html

~~Because of that, we can't remove `plugin.until.build`, and have to build Compose plugin for each new version of IDEA separately~~ (see Update)

~~Update 1:
I removed plugin.until.build anyway. In most cases there won't be a broken plugin. And if it will be broken, we will know it during IDEA EAP. Considering this, better when users don't wait the new version, and use already available one.~~

Update 2:
If we omit plugin.until.build, then it has default value +8. Returned, but made the default value big enough.

What we should do after that commit, is make sure, that we don't forget to check this compatibility. And ship a new stable version for every new stable IDEA.

The ideal way is to run `runPluginVerifier` on CI for the latest version, which will fail the build, if our plugin isn't compatible.

But for some reason, there is no EAP for IDEA 2022.3 here:
https://jb.gg/intellij-platform-builds-list