Defer dispatching coroutines resumed during animation callbacks in UI tests.
This is a workaround for the fact that we use an unconfined dispatcher
in UI tests (b/254115946), which prevents us from being able to perform
layout passes for frames in some cases (b/222093277).

TestMonotonicFrameClock gets a continuation interceptor which wraps continuations with
behavior that will usually no-op, but if the continuation is resumed
while frame callbacks are running, it will defer resuming the
underlying continuation until all the frame callbacks have finished.
This matches how continuations are dispatched in non-test code, since
that code does not use an unconfined dispatcher.

In order to preserve test delay skipping, the interceptor implements
Delay and delegates it to the underlying dispatcher, just like
ApplyingContinuationInterceptor is already doing.

This CL also adds a callback that runs after each frame in
TestMonotonicFrameClock. This change was originally in aosp/2121435.

This allows us to perform measure and layout after each frame if
necessary, which aligns the measure/layout timing with production.

Bug: b/254115946
Bug: b/222093277
Bug: b/255802670
Test: PhaseOrderingTest, TestMonotonicFrameClockTest
Relnote: "In UI tests using a Compose rule, continuations resumed during
    `withFrameNanos` callbacks will not be dispatched until after all
    frame callbacks have finished running. This matches the behavior of
    compose when running normally. However, tests that rely on the old
    behavior may fail. This should only affect code that calls
    `withFrameNanos` or `withFrameMillis` directly, and has logic
    outside of callback passed to those functions that may need to be
    moved inside the callbacks. See the animation test changes in this
    CL for examples."
Relnote: "Added optional `onPerformTraversals: (Long) -> Unit` parameter to
    `TestMonotonicFrameClock` constructor and factory function to run
    code after `withFrameNanos` callbacks but before resuming callers'
    coroutines."
Change-Id: Idb41309445a030c91e8e4ae05daa9642b450505c