Adds clipping to Modifier.pullRefreshIndicatorTransform
This ensures that when the indicator is offset outside of its bounds (i.e it has not been dragged down yet / is not refreshing), it is not visible. This fixes cases where there is no app bar to draw over the offset indicator / the indicator background color is different from the background it is drawn over.

Bug: b/248274004
Test: PullRefreshIndicatorTransformTest
Change-Id: Ic15dd35a5b3fd851bae3bc8cc6eda826e2005394