Fixes PullRefreshState not resetting indicator position back to 0
If onRefresh never ended up changing refreshing, or refreshing changed to true and false within the same frame so the state write was batched, the indicator would get stuck at its last position as we never animated it back to zero. Instead we now always animate the position back to zero, and any future change to setRefreshing will interrupt this animation and set it to the new correct position.

Fixes: b/248274004
Test: PullRefreshStateTest
Relnote: "Fixed an issue where PullRefreshIndicator could get stuck after onRefresh is called, if the refreshing state was not changed to true."
Change-Id: Ie24160bf51bd010d165d75675cfb2c51f28dde04