Introduce ability to inject a CoroutineContext into Compose UI tests.
All entry points into the Compose testing APIs – createComposeRule and
friends, and runComposeUiTest, now take an
effectContext: CoroutineContext parameter that will be used as part of
the coroutine context for composition.

Includes tests for this wiring, as well as updates two actual feature
tests that use this mechanism to inject a MotionDurationScale value to
test animation logic.

Fixes: b/265177763
Test: BasicMarqueeTest
Test: TextFieldCursorTest
Test: androidTest/ComposeUiTestTest
Test: desktopTest/ComposeUiTestTest
Relnote: "Introduced new experimental overloads for the
    `runComposeUiTest` function and `create*ComposeRule` functions that
    accept `CoroutineContext` parameters. The context will be used for
    the test composition and any `LaunchedEffect` and
    `rememberCoroutineScope()` calls in the composition."
Change-Id: I10614adabdb137ad44fb51f65403866b5b184ac1