Define ModifierNodeElements as full classes
This is an attempt to simplify some of the process of creating the
Modifier.Element object for modifiers implemented with Modifier.Node.
Using a dedicated class for the ModifierNodeElement allows us to provide
a default implementation of the layout inspector value generation which
reflectively reads all properties on the element.

As an added bonus, the `equals` and `hashCode` functions of the element
are now used to determine inequality, which can be easier to provide
since elements can be data classes, and eliminates the need for a
`key` parameter. This can also reduce allocations if a modifier needed
a compound key that would otherwise be wrapped in a new object to act
as the modifier element's key.

Relnote: """
    Removed the `modifierElementOf()` API. Please extend from
    `ModifierNodeElement` directly instead.
    """
Test: ParameterFactoryTest, all compose-ui tests
Change-Id: I2256b1776e00cbe0b781dbd7fb483e6740cdb7bc