Copy EditProcessor, EditingBuffer and EditCommand to foundation demo app.
New TextField work needs to start in foundation demo app because new text input plugin in `ui` had to be introduced in Experimental phase. Until the new plugin system becomes stable, we have to work in demo app as to not introduce any cross-experimental dependencies in foundation.

The only real change in this CL is the decoupling of EditingBuffer and EditCommand. New EditCommand is a sealed interface whose implementations should be simple data classes without any apply logic. This also makes it easy to switch the current EditingBuffer with some other data structure suitable for text buffering if need be.

Test: ./gradlew :compose:foundation:foundation:integration-tests:foundation-demos:testDebugUnitTest
Test: ./gradlew :compose:foundation:foundation:integration-tests:foundation-demos:cDAT
Change-Id: I2bdf2331b21046c6b536f536884e165178dbab15