Add option to disable compose resources generation (#4526)
# Changes

* Add `Never` to `enum class ResourceClassGeneration` to disable the
generation of Res class in the gradle plugin

# Motivation

As the [comment in issue
4229](https://github.com/JetBrains/compose-multiplatform/issues/4229#issuecomment-2008626808)
said, my team is not in the ecosystem of gradle, but organize the build
steps in bazel. We want to follow the files layout but just disable the
generation task of gradle and handle it by outself.