Add a redundancy warning to skiko.js for k/wasm target
skiko.js used to be required for k/wasm projects. But some time ago it was refactored to use skiko.mjs, so skiko.js is not needed anymore (for k/wasm, still needed for k/js). But some k/wasm  projects still include skiko.js in index.html which leads to extra loading time. The intention is to notify them without causing 404 errors for now.