Fix old version uploading to Maven Central Stage (#5185)
Fixes
https://youtrack.jetbrains.com/issue/CMP-1339/Change-description-of-androidx-artifacts-in-Maven-Central-stage
Fixes
https://youtrack.jetbrains.com/issue/CMP-7198/Maven-Central-publication-publishes-old-artifacts-for-Core-Bundle

Related change on CI:

https://jetbrains.team/p/ui/repositories/compose-teamcity-config/revision/94d34d0e

Now it matches by wildcards. It doesn't fix the core issue - associating
the exact library group with the exact artifacts, but fixing that
requires decoupling maven central upload per project instead of
comparing it by packages. It is not easy, as we not only have artifacts
in compose-multiplatform-core - we have Gradle plugin, components,
`compose-full`, Compose HTML.

## Testing
Run CI on igor.demin/upload-to-maven-central-fixes branch

Should be verified on the next 1.8.0-alpha02 release as well.