Fix typo errors in SplitPane.kt (#5200)
This PR addresses typographical errors in the comments of the
`splitpane` package. The corrections are as follows:

1. **VerticalSplitPane function comment**:
- **Original**: "Pane that place it parts **vertically** from top to
bottom and allows to change items **heights**."
- **Correction**: "Pane that places its parts **vertically** from top to
bottom and allows changing items' **heights**."

2. **HorizontalSplitPane function comment**:
- **Original**: "Pane that place it parts **horizontally** from left to
right and allows to change items **width**."
- **Correction**: "Pane that places its parts **horizontally** from left
to right and allows changing items' **widths**."

3. **SplitPane function comment**:
- **Original**: "@param isHorizontal describes is it horizontal of
vertical split pane"
- **Correction**: "@param isHorizontal describes whether it is a
horizontal or vertical split pane"

These changes enhance the clarity and grammatical accuracy of the code
comments.

**Fixes** 

N/A

**Testing**

As these changes are limited to comments and do not affect the
executable code, no additional testing is required.

**Release Notes**

N/A

Since the changes are non-functional and pertain only to code comments,
they do not necessitate an entry in the release notes.