Update Release Notes categories (#5240)
- Rename `Breaking Changes` to `Migration Notes`
- Add Lifecycle/Navigation. We used them in
[CHANGELOG.md](https://github.com/JetBrains/compose-multiplatform/blob/release/1.8.0-alpha03/CHANGELOG.md),
and have it in the script, but didn't have in the template
- Describe each category
- Use platform categories as general ones. It contradicts with my
comments
[here](https://github.com/JetBrains/compose-multiplatform/pull/5231#discussion_r1953385844)
and
[here](https://github.com/JetBrains/compose-multiplatform-core/pull/1844#discussion_r1958445125),
that it was only Compose related. But after thinking, we need some
category for changes that don't fall under Compose or any other existing
one. We can use platform categories (probably only Compose and some
Compose components fall into this category right now)

## Release Notes
N/A