KMT-599: Expose some of the preview parameters from androidx's Preview annotation (#5339)
Exposed the following Preview annotation parameters:
- name
- group
- widthDp
- heightDp
- locale
- showBackground
- backgroundColor

## Testing
Tested the local artifact in IJ: new parameters are exposed and picked
up. No additional changes on the IDE end are required.

But this should be tested by QA, of course.

## Release Notes
### Features - Multiple Platforms
- Extended the `@Preview` annotation with the following parameters:
  - name
  - group
  - widthDp
  - heightDp
  - locale
  - showBackground
  - backgroundColor

IDE (IJ or AS) will pick up these parameters in the same way it works
for `androidx` Preview annotations.