Build Helpers. Publish as a library (#5343)
Fixes
https://youtrack.jetbrains.com/issue/SKIKO-1003/Unify-Maven-publication-of-Skiko-with-Compose

This library contains logic that uploads artifacts to Maven Central.
[Previously](https://github.com/JetBrains/compose-multiplatform/pull/5184)
it was converted to CLI-only. But now we convert it back to a library,
because it is currently [the best way to use
it](https://youtrack.jetbrains.com/issue/SKIKO-1003/Unify-Maven-publication-of-Skiko-with-Compose#focus=Comments-27-12193419.0-0)
from Skiko. We keep the CLI as well, it is used for Compose/Androidx
currently

On every change in the source, CI will publish
org.jetbrains.compose.internal.build-helpers:publishing:$version

[Skiko PR](https://github.com/JetBrains/skiko/pull/1067)
[TeamCity auto-publishing
configuration](https://jetbrains.team/p/ui/repositories/compose-teamcity-config/files/igor.demin/publish-build-helpers/.teamcity/compose/service/PublishBuildHelpers.kt)
(will be merged after this PR)

## Testing
- `./gradlew publishToMavenLocal` and use it in Skiko
- run CLI:
```
./gradlew -p=cli reuploadArtifactsToMavenCentral -Pmaven.central.sign=true \
-Pmaven.central.coordinates=org.jetbrains.compose*:*:%version.COMPOSE%,org.jetbrains.compose.material:material-navigation*:%version.COMPOSE_MATERIAL_NAVIGATION% \
-Pmaven.central.stage=org.jetbrains.compose \
-Pmaven.central.description="Compose %version.COMPOSE% and associated libs" \
-Pmaven.central.staging.close.after.upload=true
```
- will battle test it after merging, trying to publish Skiko to Maven
Central

## Release Notes
N/A