Use Web Cache API for extra layer of resources caching (#5379)
Use Web Cache API for all resources 

The Cache is reset on every app launch (page refresh).
The initial idea was to reset the Cache only when a new session starts,
but we risk to have an outdated resources state (it can be incompatible
with the app logic expectations and lead to crashes).

Fixes https://youtrack.jetbrains.com/issue/CMP-7996 

## Testing
This should be tested by QA

## Release Notes
### Fixes - Resources
- Use Web Cache API for all resources to avoid repeated and redundant
HTTP requests