CMP-8068 Fix crash in `getSystemEnvironment` on headless desktop (#5471)
Fixes a crash in `getSystemEnvironment` on desktop when running in a
headless environment (such as most CI runners) by providing a default
density qualifier. In the future that might be something that's nice to
make configurable, but since both our use case and the linked issue
encountered it when calling `getString` I didn't worry about it too
hard.

Fixes [CMP-8068](https://youtrack.jetbrains.com/issue/CMP-8068)

## Testing

Made a local build and used it to run some tests in another project that
were failing on calls to `getString`

## Release Notes
### Fixes - Resources
- Fix crash in `getSystemEnvironment` on headless desktop